# openapi.yaml
openapi: 3.0.3
info:
  title: E-ES Link API
  description: |
    API for the E-ES Link championship and team application platform.
    This specification describes both implemented endpoints and planned operations.
  version: 1.1.0
  contact:
    name: E-ES Link Team
servers:
  - url: https://izv.svaza.lv
    description: Production server
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /api/submit-form:
    post:
      tags:
        - Test
      summary: Submit test form data
      description: Internal endpoint for test form submission.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Received form data
  /api/register:
    post:
      tags:
        - Accounts
      summary: Register a new account or submit a new account application
      description: |
        Registers a new account directly for applications without a team,
        or submits an account application for verification when the request is team-related.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountApplication'
      responses:
        '200':
          description: Accepted or pending verification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          description: Invalid request
  /api/login:
    get:
      tags:
        - Auth
      summary: Get current authenticated account
      description: Returns the authenticated admin or user account from the current request context.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Authenticated account information
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/AccountInfo'
                  - $ref: '#/components/schemas/AdminInfo'
        '401':
          description: Unauthorized
  /api/account-applications:
    get:
      tags:
        - Accounts
      summary: List pending account applications
      description: |
        Returns all account applications for superadmins, or team-specific applications for team leaders.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of account applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountApplication'
        '403':
          description: Forbidden
  /api/account-verification-info/{key}:
    get:
      tags:
        - Accounts
      summary: Get verification criteria for an account application
      description: Returns verification metadata for a specific account application.
      security:
        - BearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: Account application key
      responses:
        '200':
          description: Verification criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountVerificationCriteria'
        '404':
          description: Application not found
        '401':
          description: Unauthorized
  /api/verify-account-application/{key}:
    post:
      tags:
        - Accounts
      summary: Verify and register an account application
      description: |
        Verifies the account application identified by key and registers the account.
        Superadmins can verify team leader applications and all other applications.
      security:
        - BearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: Account application key
      responses:
        '200':
          description: Account verified and registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '403':
          description: Forbidden
        '404':
          description: Application not found
...
  /api/accounts:
    get:
      tags:
        - Accounts
      summary: List registered accounts
      description: Returns all registered accounts for superadmins, or team-specific accounts for team leaders.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Accounts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountInfo'
        '403':
          description: Forbidden
  /api/account/{key}:
    patch:
      tags:
        - Accounts
      summary: Update a registered account
      description: Modify an existing registered account by key (superadmin only).
      security:
        - BearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: Account key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountInfo'
      responses:
        '200':
          description: Account updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '403':
          description: Forbidden
        '404':
          description: Account not found
  /api/admin-accounts:
    get:
      tags:
        - Admin
      summary: List admin accounts
      description: Return all admin accounts (superadmin only).
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Admin accounts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminInfo'
        '403':
          description: Forbidden
  /api/admin-account/{key}:
    patch:
      tags:
        - Admin
      summary: Update an admin account
      description: Modify an existing admin account by key (superadmin only).
      security:
        - BearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: Admin key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminInfo'
      responses:
        '200':
          description: Admin account updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '403':
          description: Forbidden
        '404':
          description: Admin account not found
  /api/race/start:
    post:
      tags:
        - Races
      summary: Start a race for a competitor
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RaceStartRequest'
      responses:
        '200':
          description: Race started successfully
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
  /api/car/finish:
    post:
      tags:
        - Races
      summary: Finish a race for a competitor
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarFinishRequest'
      responses:
        '200':
          description: Race finished successfully
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
  /api/races:
    get:
      tags:
        - Races
      summary: Get all race definitions
      responses:
        '200':
          description: Race configurations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RaceLap'
    post:
      tags:
        - Races
      summary: Update race configurations
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/RaceLap'
      responses:
        '200':
          description: Race configurations saved
        '401':
          description: Unauthorized
...
  /api/results/{raceName}:
    get:
      tags:
        - Results
      summary: Get race results
      parameters:
        - name: raceName
          in: path
          required: true
          schema:
            type: string
          description: Race name
      responses:
        '200':
          description: Race results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RaceResult'
        '404':
          description: Race not found
  /api/points:
    post:
      tags:
        - Points
      summary: Submit points for competitors
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PointsSubmission'
      responses:
        '200':
          description: Points saved successfully
        '401':
          description: Unauthorized
    delete:
      tags:
        - Points
      summary: Delete points for a race
      security:
        - BearerAuth: []
      parameters:
        - name: raceName
          in: query
          required: true
          schema:
            type: string
          description: Name of the race to clear points for
      responses:
        '200':
          description: Points deleted successfully
        '401':
          description: Unauthorized
  /api/leaderboard/{ageGroup}:
    get:
      tags:
        - Leaderboard
      summary: Get leaderboard by age group
      parameters:
        - name: ageGroup
          in: path
          required: true
          schema:
            type: string
          description: Age group identifier or "all"
      responses:
        '200':
          description: Leaderboard data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardEntry'
  /api/news:
    get:
      tags:
        - News
      summary: Get all news articles
      responses:
        '200':
          description: News articles list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NewsArticle'
    post:
      tags:
        - News
      summary: Create a news article
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsArticle'
      responses:
        '201':
          description: Article created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsArticle'
        '401':
          description: Unauthorized
  /api/news/{id}:
    get:
      tags:
        - News
      summary: Get a news article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: News article
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsArticle'
        '404':
          description: Not found
    put:
      tags:
        - News
      summary: Update a news article
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsArticle'
      responses:
        '200':
          description: Article updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsArticle'
        '404':
          description: Not found
        '401':
          description: Unauthorized
    delete:
      tags:
        - News
      summary: Delete a news article
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Article deleted successfully
        '404':
          description: Not found
        '401':
          description: Unauthorized
  /api/news/{id}/attachments:
    post:
      tags:
        - News
      summary: Upload attachment to a news article
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Attachment uploaded successfully
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '404':
          description: Article not found
  /api/events:
    get:
      tags:
        - Events
      summary: Get all events
      responses:
        '200':
          description: Event list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompetitionEvent'
    post:
      tags:
        - Events
      summary: Create an event
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompetitionEvent'
      responses:
        '201':
          description: Event created successfully
        '401':
          description: Unauthorized
  /api/events/{id}:
    get:
      tags:
        - Events
      summary: Get a specific event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitionEvent'
        '404':
          description: Not found
    put:
      tags:
        - Events
      summary: Update an event
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompetitionEvent'
      responses:
        '200':
          description: Event updated successfully
        '404':
          description: Not found
        '401':
          description: Unauthorized
    delete:
      tags:
        - Events
      summary: Delete an event
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event deleted successfully
        '404':
          description: Not found
        '401':
          description: Unauthorized
...
  /api/team-application:
    post:
      tags:
        - Teams
      summary: Create a team application
      description: Submit a new team registration application.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamApplication'
      responses:
        '201':
          description: Team application created
        '400':
          description: Invalid request
  /api/team:
    patch:
      tags:
        - Teams
      summary: Update team application data by key
      description: Update a team application using data in the request body.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamApplicationPatch'
      responses:
        '200':
          description: Team application updated
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
  /api/team/{key}:
    post:
      tags:
        - Teams
      summary: Update team application data by key path
      description: Submit updated team data using the application key in the path.
      security:
        - BearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamApplicationPatch'
      responses:
        '200':
          description: Team application updated
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
  /ws:
    get:
      tags:
        - WebSocket
      summary: WebSocket live data endpoint
      description: Upgrade to a WebSocket connection for live tracking data.
      parameters:
        - name: token
          in: query
          required: false
          schema:
            type: string
          description: Optional access token
      responses:
        '101':
          description: WebSocket upgrade successful
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token for authenticated API access.
  schemas:
    StatusResponse:
      type: object
      properties:
        status:
          type: string
          example: accepted
    AccountApplication:
      type: object
      required:
        - fullName
        - dateOfBirth
        - parole
        - lietotajvards
        - epasts
        - telefonanumurs
        - komandasNosaukums
        - role
      properties:
        key:
          type: string
          description: Application key, assigned dynamically
        fullName:
          type: string
          example: Janis Berzins
        dateOfBirth:
          type: string
          format: date
          example: 2005-04-12
        parole:
          type: string
          example: secret123
        lietotajvards:
          type: string
          example: janis.b
        epasts:
          type: string
          format: email
          example: janis@example.com
        telefonanumurs:
          type: string
          example: +37112345678
        komandasNosaukums:
          type: string
          example: Volt Racers
        role:
          type: string
          enum: [team_leader, member]
          example: member
    TeamMember:
      type: object
      properties:
        key:
          type: string
        fullName:
          type: string
        dateOfBirth:
          type: string
          format: date
        educationalInstitution:
          type: string
        role:
          type: string
          example: member
        classOrYear:
          type: string
        id:
          type: string
    ResponsiblePerson:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        status:
          type: string
    AccountInfo:
      type: object
      properties:
        dati:
          $ref: '#/components/schemas/TeamMember'
        lietotajvards:
          type: string
        epasts:
          type: string
          format: email
        telefonanumurs:
          type: string
    AdminInfo:
      type: object
      properties:
        key:
          type: string
        lietotajvards:
          type: string
        superadmin:
          type: boolean
    AccountVerificationCriteria:
      type: object
      properties:
        canRegister:
          type: boolean
        requiredFieldsPresent:
          type: boolean
        usernameAvailable:
          type: boolean
        noTeamNameProvided:
          type: boolean
        notTeamLeaderRole:
          type: boolean
        teamNameProvided:
          type: boolean
        teamApplicationExists:
          type: boolean
        teamMemberFound:
          type: boolean
        teamMemberRoleMatches:
          type: boolean
        noMatchingTeamMember:
          type: boolean
        teamLeaderRole:
          type: boolean
        teamLeaderMemberFound:
          type: boolean
        teamLeaderDateOfBirthMatch:
          type: boolean
    TeamApplication:
      type: object
      required:
        - teamName
        - ageGroup
        - cityOrRegion
        - members
        - responsiblePerson
      properties:
        id:
          type: string
        teamName:
          type: string
          example: Volt Racers
        ageGroup:
          type: string
          example: A
        institution:
          type: string
          example: Rīgas Valsts 1. ģimnāzija
        cityOrRegion:
          type: string
          example: Riga
        members:
          type: array
          items:
            $ref: '#/components/schemas/TeamMember'
        responsiblePerson:
          $ref: '#/components/schemas/ResponsiblePerson'
        howHeardAbout:
          type: string
        comments:
          type: string
        confirmTruthful:
          type: boolean
        confirmRules:
          type: boolean
        confirmMedia:
          type: boolean
        appliedAt:
          type: string
          format: date-time
        status:
          type: string
          example: pending
    TeamApplicationPatch:
      type: object
      properties:
        key:
          type: string
        teamName:
          type: string
        ageGroup:
          type: string
        institution:
          type: string
        cityOrRegion:
          type: string
        members:
          type: array
          items:
            $ref: '#/components/schemas/TeamMember'
        responsiblePerson:
          $ref: '#/components/schemas/ResponsiblePerson'
        howHeardAbout:
          type: string
        comments:
          type: string
        status:
          type: string
    RaceStartRequest:
      type: object
      required:
        - raceName
        - ID
      properties:
        raceName:
          type: string
        ID:
          type: integer
    CarFinishRequest:
      type: object
      required:
        - ID
      properties:
        ID:
          type: integer
    RaceLap:
      type: object
      required:
        - RaceName
        - Lap
        - Length
      properties:
        RaceName:
          type: string
        Lap:
          type: integer
        Length:
          type: number
          format: float
...
    RaceResult:
      type: object
      properties:
        ID:
          type: integer
        Username:
          type: string
        avatar:
          type: string
        RaceName:
          type: string
        Lap:
          type: integer
        usedEnergy:
          type: number
        efficiency:
          type: number
        shellEfficiency:
          type: number
        averagePower:
          type: number
        averageSpeed:
          type: number
        elapsedTime:
          type: string
        points:
          type: number
    PointsSubmission:
      type: object
      required:
        - raceName
        - points
      properties:
        raceName:
          type: string
        points:
          type: array
          items:
            type: object
            properties:
              ID:
                type: integer
              points:
                type: number
    LeaderboardEntry:
      type: object
      properties:
        ID:
          type: integer
        username:
          type: string
        avatar:
          type: string
        position_change:
          type: integer
        categoryNames:
          type: array
          items:
            type: string
        categoryPoints:
          type: array
          items:
            type: number
    NewsArticle:
      type: object
      required:
        - id
        - title
        - summary
        - content
        - author
        - date
        - tags
        - published
      properties:
        id:
          type: string
        title:
          $ref: '#/components/schemas/BilingualText'
        summary:
          $ref: '#/components/schemas/BilingualText'
        content:
          $ref: '#/components/schemas/BilingualText'
        author:
          type: string
        date:
          type: string
          format: date
        image:
          type: string
          format: uri
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/AttachedFile'
    AttachedFile:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        url:
          type: string
          format: uri
    CompetitionEvent:
      type: object
      required:
        - id
        - title
        - description
        - date
        - location
        - status
      properties:
        id:
          type: string
        title:
          $ref: '#/components/schemas/BilingualText'
        description:
          $ref: '#/components/schemas/BilingualText'
        date:
          type: string
          format: date
        endDate:
          type: string
          format: date
        location:
          $ref: '#/components/schemas/BilingualText'
        registrationUrl:
          type: string
          format: uri
        status:
          type: string
          enum: [upcoming, ongoing, completed]
    BilingualText:
      type: object
      required:
        - en
        - lv
      properties:
        en:
          type: string
        lv:
          type: string